% File src/library/grDevices/man/cm.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{cm}
\alias{cm}
\title{Unit Transformation}
\description{
  Translates from inches to cm (centimeters).
}
\usage{
cm(x)
}
\arguments{
  \item{x}{numeric vector}
}
\examples{
cm(1)# = 2.54

}
\keyword{dplot}
