% File src/library/grDevices/man/grDevices-internal.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2013 R Core Team
% Distributed under GPL 2 or later

\name{grDevices-internal}
\alias{.clipPath}
\alias{.defineGroup}
\alias{.devUp}
\alias{.linearGradientPattern}
\alias{.mask}
\alias{.opIndex}
\alias{.radialGradientPattern}
\alias{.ruleIndex}
\alias{.setClipPath}
\alias{.setMask}
\alias{.setPattern}
\alias{.tilingPattern}
\alias{.useGroup}
\title{Internal grDevices Functions}
\description{
  Internal grDevices functions
}
\details{
  These are used for communication between the graphics engine and
  graphics systems, \pkg{graphics} and \pkg{grid}, and are
  not to be called by the user.
}
\keyword{ internal }
