% File src/library/grDevices/man/grDevices-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2009 R Core Team
% Distributed under GPL 2 or later

\name{grDevices-package}
\alias{grDevices-package}
\alias{grDevices}
\docType{package}
\title{
  The R Graphics Devices and Support for Colours and Fonts
}
\description{
  Graphics devices and support for base and grid graphics
}
\details{

  This package contains functions which support both
  \link[=graphics-package]{base} and \link[=grid-package]{grid} graphics.

  For a complete list of functions, use \code{library(help = "grDevices")}.

}
\author{
  R Core Team and contributors worldwide

  Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ device }
