\name{grDevices-defunct}
\alias{grDevices-defunct}
#ifdef unix
\alias{GNOME}
\alias{gnome}
#endif
\title{Defunct Functions in Package grDevices}
\description{
  The functions or variables listed here are no longer part of \R as
  they are not needed (any more).
}
#ifdef unix
\usage{
gnome(display="", width=7, height=7, pointsize=12)
GNOME(display="", width=7, height=7, pointsize=12)
}
#endif
\seealso{
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{internal}
